requirements.txt
customtkinter
opencv-python
Pillow
pygame
pytesseract
gTTS
tensorflow
